pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gotham"
version = "0.1.0"
description = "A crime management API built with Robyn and PostgreSQL"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Paul Roujansky", email = "paul@roujansky.eu" }
]
dependencies = [
    "robyn>=0.37.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.2",
]

[dependency-groups]
dev = [
    "ruff>=0.11.10",
]

[project.urls]
Homepage = "https://github.com/paulroujansky/gotham"
Documentation = "https://github.com/paulroujansky/gotham#readme"
Repository = "https://github.com/paulroujansky/gotham.git"
Issues = "https://github.com/paulroujansky/gotham/issues"

[tool.ruff]
line-length = 88
target-version = "py311"
select = ["E", "F", "B", "I"]
ignore = []

[tool.ruff.per-file-ignores]
"src/gotham/core/crud.py" = ["E501"]  # E501 is the code for line too long
